Move generic parts out of autobahn and fix relative imports
