+++ src/roslibpy/comm/comm_autobahn.py
@@ -6,75 +6,15 @@ import logging
 from autobahn.twisted.websocket import WebSocketClientFactory, WebSocketClientProtocol
 from twisted.internet.protocol import ReconnectingClientFactory
 
-from . import Message, ServiceResponse
-from .event_emitter import EventEmitterMixin
+from . import RosBridgeProtocol, RosBridgeException
+from .. import Message, ServiceResponse
+from ..event_emitter import EventEmitterMixin
 
 LOGGER = logging.getLogger('roslibpy')
 
-
-class RosBridgeException(Exception):
-    """Exception raised on the ROS bridge communication."""
-    pass
-
-
-class RosBridgeProtocol(WebSocketClientProtocol):
-    """Implements the websocket client protocol to encode/decode JSON ROS Bridge messages."""
-
+class AutobahnRosBridgeProtocol(RosBridgeProtocol, WebSocketClientProtocol):
     def __init__(self, *args, **kwargs):
-        super(RosBridgeProtocol, self).__init__(*args, **kwargs)
-        self.factory = None
-        self._pending_service_requests = {}
-        self._message_handlers = {
-            'publish': self._handle_publish,
-            'service_response': self._handle_service_response,
-            'call_service': self._handle_service_request,
-        }
-        # TODO: add handlers for op: status
-
-    def send_ros_message(self, message):
-        """Encode and serialize ROS Bridge protocol message.
-
-        Args:
-            message (:class:`.Message`): ROS Bridge Message to send.
-        """
-        try:
-            json_message = json.dumps(dict(message)).encode('utf8')
-            LOGGER.debug('Sending ROS message: %s', json_message)
-
-            self.sendMessage(json_message)
-        except Exception as exception:
-            # TODO: Check if it makes sense to raise exception again here
-            # Since this is wrapped in many layers of indirection
-            LOGGER.exception('Failed to send message, %s', exception)
-
-    def register_message_handlers(self, operation, handler):
-        """Register a message handler for a specific operation type.
-
-        Args:
-            operation (:obj:`str`): ROS Bridge operation.
-            handler: Callback to handle the message.
-        """
-        if operation in self._message_handlers:
-            raise RosBridgeException(
-                'Only one handler can be registered per operation')
-
-        self._message_handlers[operation] = handler
-
-    def send_ros_service_request(self, message, callback, errback):
-        """Initiate a ROS service request through the ROS Bridge.
-
-        Args:
-            message (:class:`.Message`): ROS Bridge Message containing the service request.
-            callback: Callback invoked on successful execution.
-            errback: Callback invoked on error.
-        """
-        request_id = message['id']
-        self._pending_service_requests[request_id] = (callback, errback)
-
-        json_message = json.dumps(dict(message)).encode('utf8')
-        LOGGER.debug('Sending ROS service request: %s', json_message)
-
-        self.sendMessage(json_message)
+        super(AutobahnRosBridgeProtocol, self).__init__(*args, **kwargs)
 
     def onConnect(self, response):
         LOGGER.debug('Server connected: %s', response.peer)
@@ -95,44 +35,21 @@ class RosBridgeProtocol(WebSocketClientProtocol):
 
         handler(message)
 
-    def _handle_publish(self, message):
-        self.factory.emit(message['topic'], message['msg'])
-
-    def _handle_service_response(self, message):
-        request_id = message['id']
-        service_handlers = self._pending_service_requests.get(request_id, None)
-
-        if not service_handlers:
-            raise RosBridgeException(
-                'No handler registered for service request ID: "%s"' % request_id)
-
-        callback, errback = service_handlers
-        del self._pending_service_requests[request_id]
-
-        if 'result' in message and message['result'] is False:
-            if errback:
-                errback(message['values'])
-        else:
-            if callback:
-                callback(ServiceResponse(message['values']))
-
-    def _handle_service_request(self, message):
-        if 'service' not in message:
-            raise ValueError(
-                'Expected service name missing in service request')
-
-        self.factory.emit(message['service'], message)
-
     def onClose(self, wasClean, code, reason):
         LOGGER.info('WebSocket connection closed: %s', reason)
 
+    def send_message(self, payload):
+        return self.sendMessage(payload, isBinary=False, fragmentSize=None, sync=False, doNotCompress=False)
 
-class RosBridgeClientFactory(EventEmitterMixin, ReconnectingClientFactory, WebSocketClientFactory):
+    def send_close(self):
+        self.sendClose()
+
+class AutobahnRosBridgeClientFactory(EventEmitterMixin, ReconnectingClientFactory, WebSocketClientFactory):
     """Factory to construct instance of the ROS Bridge protocol."""
-    protocol = RosBridgeProtocol
+    protocol = AutobahnRosBridgeProtocol
 
     def __init__(self, *args, **kwargs):
-        super(RosBridgeClientFactory, self).__init__(*args, **kwargs)
+        super(AutobahnRosBridgeClientFactory, self).__init__(*args, **kwargs)
         self._proto = None
         self.setProtocolOptions(closeHandshakeTimeout=5)
